feat(frontend): add GitHub Pages static export support (#1)
- Configure next.config.js for static export with basePath
- Add GitHub Actions workflow for automatic deployment
- Split pages into server/client components for generateStaticParams
- Add out/ to gitignore
