+++ frontend/next.config.js
@@ -3,15 +3,22 @@ const nextConfig = {
   // Enable React strict mode for development
   reactStrictMode: true,
 
-  // API proxy for development
-  async rewrites() {
-    return [
-      {
-        source: '/api/v1/:path*',
-        destination: `${process.env.NEXT_PUBLIC_API_URL || 'http://localhost:8000/api/v1'}/:path*`,
-      },
-    ];
+  // Static export for GitHub Pages
+  output: 'export',
+
+  // Base path for GitHub Pages (repository name)
+  basePath: process.env.NODE_ENV === 'production' ? '/hr-system-agent' : '',
+
+  // Asset prefix for GitHub Pages
+  assetPrefix: process.env.NODE_ENV === 'production' ? '/hr-system-agent/' : '',
+
+  // Disable image optimization (requires server)
+  images: {
+    unoptimized: true,
   },
+
+  // Trailing slash for static export
+  trailingSlash: true,
 };
 
 module.exports = nextConfig;
